Update converter separation by data version
